No concept of test namespace anymore
