+++ test/testdata/packager/package-prefix-enforcement/critic_prefix/test/__package.rb
@@ -1,0 +1,6 @@
+# typed: strict
+
+class Test::Critic::SomePkg < PackageSpec
+  test!
+  import Critic::SomePkg, uses_internals: true
+end
