+++ test/testdata/packager/package-prefix-enforcement/nested/test/__package.rb
@@ -1,0 +1,6 @@
+# typed: strict
+
+class Test::Root::Nested < PackageSpec
+  test!
+  import Root::Nested, uses_internals: true
+end
